Корректировка для прохождения тестов №1
