+++ js/form.js
@@ -1,34 +1,9 @@
 import {sendData} from './api.js';
 import {showUploadError, showUploadSuccess} from './alerts.js';
 
-const imgUploadForm = document.querySelector('.img-upload__form');
+export const imgUploadForm = document.querySelector('.img-upload__form');
 export const imgUpload = imgUploadForm.querySelector('.img-upload__input');
-const imgUploadEffects = imgUploadForm.querySelector('.img-upload__effects');
-const effectsNodeList = imgUploadEffects.querySelectorAll('.effects__item');
-const uploadedImgPreview = imgUploadForm.querySelector('.img-upload__preview');
-const FILE_TYPES = ['jpg', 'jpeg', 'png'];
-
-imgUpload.value = '';
-
-// Загрузка изображения в форму
-
-export const imgUploadFormRender = () => {
-  const loadedImg = imgUpload.files[0];
-  const fileName = loadedImg.name.toLowerCase();
-  const matches = FILE_TYPES.some((it) => fileName.endsWith(it));
-
-  uploadedImgPreview.querySelector('img').src = '';
-
-  if (matches) {
-    const loadedImgUrl = URL.createObjectURL(loadedImg);
-
-    uploadedImgPreview.querySelector('img').src = URL.createObjectURL(loadedImg);
-
-    for (let i = 0; i < effectsNodeList.length; i++) {
-      effectsNodeList[i].querySelector('.effects__preview').style.backgroundImage = `url(${loadedImgUrl})`;
-    }
-  }
-};
+export const uploadedImgPreview = imgUploadForm.querySelector('.img-upload__preview');
 
 // Валидация формы
 
@@ -124,172 +99,5 @@ export const setUserPhotoSubmit = (onSuccess) => {
   });
 };
 
-
 pristine.reset();
 pristine.destroy();
-
-export const cleanForm = () => {
-  uploadedImgPreview.querySelector('img').src = '';
-  imgUpload.value = '';
-  imgUploadForm.querySelector('.text__hashtags').value = '';
-  imgUploadForm.querySelector('.text__description').value = '';
-};
-
-// Масштабирование загруженного изображения
-
-const scaleControl = imgUploadForm.querySelector('.scale__control--value');
-const scaleBiggerButton = imgUploadForm.querySelector('.scale__control--bigger');
-const scaleSmallerButton = imgUploadForm.querySelector('.scale__control--smaller');
-let scaleCounter = parseInt(scaleControl.value, 10);
-
-const changeScale = () => {
-  const scale = scaleCounter / 100;
-  scaleControl.setAttribute('value', `${scaleCounter}%`);
-  uploadedImgPreview.style.transform = `scale(${scale})`;
-};
-
-scaleBiggerButton.addEventListener('click', (evt) => {
-  evt.preventDefault();
-  if (scaleCounter < 100) {
-    scaleCounter += 25;
-    changeScale();
-  }
-});
-
-scaleSmallerButton.addEventListener('click', (evt) => {
-  evt.preventDefault();
-  if (scaleCounter > 25) {
-    scaleCounter -= 25;
-    changeScale();
-  }
-});
-
-// Эффекты для изображения
-
-const sliderContainer = imgUploadForm.querySelector('.img-upload__effect-level');
-const sliderLevelInput = imgUploadForm.querySelector('.effect-level__value');
-const sliderLevelCount = imgUploadForm.querySelector('.effect-level__value');
-const sliderElement = imgUploadForm.querySelector('.effect-level__slider');
-const effectsRadios = imgUploadForm.querySelectorAll('.effects__radio');
-const originalEffect = imgUploadForm.querySelector('#effect-none');
-const chromeEffect = imgUploadForm.querySelector('#effect-chrome');
-const sepiaEffect = imgUploadForm.querySelector('#effect-sepia');
-const marvinEffect = imgUploadForm.querySelector('#effect-marvin');
-const phobosEffect = imgUploadForm.querySelector('#effect-phobos');
-const heatEffect = imgUploadForm.querySelector('#effect-heat');
-
-noUiSlider.create(sliderElement, {
-  start: 1,
-  connect: true,
-  range: {
-    'min': 0,
-    'max': 1,
-  },
-  step: 0.1,
-});
-
-sliderElement.noUiSlider.on('update', () => {
-  sliderLevelInput.value = sliderElement.noUiSlider.get();
-  sliderLevelCount.textContent = sliderLevelInput.value;
-});
-
-if (originalEffect.checked) {
-  sliderContainer.classList.add('hidden');
-}
-
-for (const effectRadio of effectsRadios) {
-  effectRadio.addEventListener('change',() => {
-    if (originalEffect.checked) {
-      sliderContainer.classList.add('hidden');
-
-      uploadedImgPreview.style.filter = '';
-    }
-
-    if (chromeEffect.checked) {
-      sliderContainer.classList.remove('hidden');
-
-      sliderElement.noUiSlider.updateOptions({
-        start: 1,
-        range: {
-          'min': 0,
-          'max': 1,
-        },
-        step: 0.1,
-      });
-
-      sliderElement.noUiSlider.on('update', () => {
-        uploadedImgPreview.style.filter = `grayscale(${sliderLevelInput.value})`;
-      });
-    }
-
-    if (sepiaEffect.checked) {
-      sliderContainer.classList.remove('hidden');
-
-      sliderElement.noUiSlider.updateOptions({
-        start: 1,
-        range: {
-          'min': 0,
-          'max': 1,
-        },
-        step: 0.1,
-      });
-
-      sliderElement.noUiSlider.on('update', () => {
-        uploadedImgPreview.style.filter = `sepia(${sliderLevelInput.value})`;
-      });
-    }
-
-    if (marvinEffect.checked) {
-      sliderContainer.classList.remove('hidden');
-
-      sliderElement.noUiSlider.updateOptions({
-        start: 100,
-        range: {
-          'min': 0,
-          'max': 100,
-        },
-        step: 1,
-      });
-
-      sliderElement.noUiSlider.on('update', () => {
-        uploadedImgPreview.style.filter = `invert(${sliderLevelInput.value}%)`;
-      });
-    }
-
-    if (phobosEffect.checked) {
-      sliderContainer.classList.remove('hidden');
-
-      sliderElement.noUiSlider.updateOptions({
-        start: 3,
-        range: {
-          'min': 0,
-          'max': 3,
-        },
-        step: 0.1,
-      });
-
-      sliderElement.noUiSlider.on('update', () => {
-        uploadedImgPreview.style.filter = `blur(${sliderLevelInput.value}px)`;
-      });
-    }
-
-    if (heatEffect.checked) {
-      sliderContainer.classList.remove('hidden');
-
-      sliderElement.noUiSlider.updateOptions({
-        start: 3,
-        range: {
-          'min': 1,
-          'max': 3,
-        },
-        step: 0.1,
-      });
-
-      sliderElement.noUiSlider.on('update', () => {
-        uploadedImgPreview.style.filter = `brightness(${sliderLevelInput.value})`;
-      });
-    }
-  });
-}
-
-
